src/tiny_chatbot/main.py
import os
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI

# Load environment variables from .env file
load_dotenv()

def main():
    llm = ChatOpenAI(model="gpt-4o-mini", temperature=0)
    response = llm.invoke("Say hello from tiny-chatbot!")
    print(response.content)

if __name__ == "__main__":
    main()