pyproject.toml
[tool.poetry]
name = "tiny-chatbot"
version = "0.1.0"
description = "A tiny chatbot project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tiny_chatbot"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
langchain = "^1.1.0"
langchain-core = "^1.1.0"
langchain-community = "^0.4.1"
openai = "^2.8.1"
anthropic = "^0.75.0"
langchain-openai = "^1.1.0"
langchain-anthropic = "^1.2.0"
langgraph = "^1.0.4"
faiss-cpu = "^1.13.0"
fastapi = "^0.122.0"
uvicorn = "^0.38.0"
pypdf = "^6.4.0"
python-dotenv = "^1.2.1"
mcp = "^1.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

